fix: refactor repr method and changed filename
